Simplify API to only have category/project instead of 3-levels (#6)
A category is like a grouping of related projects, and a project is what
is now assoiated with a specific time-record.  This also makes the
concept of the project slug globally unique, and removes the slug
property from the category.  Also the API has been refactored such that
the project URL doesn't depend on the category PK, as the category is
now passed as part of the JSON.