Prevent deleting categories and projects with records
This adds the number of records to the internal object representations
for both projects and categories and disables the "delete" menu option
on any project or category with records.  Also the number of records is
now displayed in the table, making it easier to determine which projects
and categories are safe to clean up.

fixes #70

Report number of associated records on each project

Disable the "delete" menu on projects with existing records

Show number of records in project table and allow sorting

Add the number of records to the category views

Add the number of records to the category table

Signed-off-by: Anthony Oteri <anthony.oteri@edgeware.tv>